Made som general fixes

+++ core/src/main/java/com/g16/feyrune/model/combat/AttackHandler.java
@@ -9,12 +9,13 @@ public class AttackHandler {
 
     /**
      * Used to make a monster attack another monster
+     *
      * @param attacker The monster performing the attack
      * @param defender The monster trying to evade the attack
-     * @param attack The specific attack being used
+     * @param attack   The specific attack being used
      */
-    public static void handleAttack(ICombatCreature attacker, ICombatCreature defender, IAbility attack){
-        if (evasiveManoeuvre(defender, attack)){
+    public static void handleAttack(ICombatCreature attacker, ICombatCreature defender, IAbility attack) {
+        if (evasiveManoeuvre(defender, attack)) {
             int damage = calculateDamage(attacker, defender, attack);
             defender.takeDamage(damage);
         }
@@ -22,25 +23,28 @@ public class AttackHandler {
 
     /**
      * Calculates the damage done by an attack
+     *
      * @param attacker: creature containing the attack used in the action
-     * @param attack: the attack action used against the defender
+     * @param attack:   the attack action used against the defender
      * @return damage dealt to the defender (only calculates, doesn't deal it)
      */
-    private static int calculateDamage(ICombatCreature attacker,ICombatCreature defender, IAbility attack){
-        int damage = (int) ((double)(attacker.getPower() * ((double)attack.getAttackPower()/100))/defender.getDefense());
+    private static int calculateDamage(ICombatCreature attacker, ICombatCreature defender, IAbility attack) {
+        int damage = (int) ((attacker.getPower() * ((double) attack.getAttackPower() / 100)) / defender.getDefense());
         return damage;
     }
 
     /**
-     * Calculates whether the attack accuracy is high enough to hit and whether the defender's evasiveness is high enough to evade it
-     * @param defender: monster trying to evade the attack
+     * Calculates whether the attack accuracy is high enough to hit and whether
+     * the defender's evasiveness is high enough to evade it
+     *
+     * @param defender:    monster trying to evade the attack
      * @param baseAbility: attack to check accuracy on.
      * @return if attack is successful: true, if it misses false
      */
-    private static boolean evasiveManoeuvre(ICombatCreature defender, IAbility baseAbility){
-        boolean evasion = defender.getSpeed()< Random.randomInt(100);
-        boolean accuracy = baseAbility.getAttackAccuracy()>Random.randomInt(100);
-        boolean hit= evasion && accuracy;
+    private static boolean evasiveManoeuvre(ICombatCreature defender, IAbility baseAbility) {
+        boolean evasion = defender.getSpeed() < Random.randomInt(100);
+        boolean accuracy = baseAbility.getAttackAccuracy() > Random.randomInt(100);
+        boolean hit = evasion && accuracy;
         return hit;
     }
 }
